pyproject.toml
[project]
name = "agenticioc"
version = "0.1.0"
description = "IoC scaffold for building hierarchical autonomous agents in a modular and pluggable way"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
